core: paths: move envelope train path
