+++ core/envelope-sim/src/test/kotlin/fr/sncf/osrd/envelope_sim/EnvelopeSimPathTest.kt
@@ -4,10 +4,10 @@ import com.google.common.collect.ImmutableRangeMap
 import com.google.common.collect.Range
 import com.google.common.collect.RangeMap
 import com.google.common.collect.TreeRangeMap
-import fr.sncf.osrd.envelope_sim.electrification.Electrified
-import fr.sncf.osrd.envelope_sim.electrification.NonElectrified
 import fr.sncf.osrd.path.implementations.EnvelopeSimPath
 import fr.sncf.osrd.path.interfaces.Electrification
+import fr.sncf.osrd.path.legacy_objects.electrification.Electrified
+import fr.sncf.osrd.path.legacy_objects.electrification.NonElectrified
 import fr.sncf.osrd.utils.RangeMapUtils
 import org.junit.jupiter.api.Assertions
 import org.junit.jupiter.api.Test
