+++ core/envelope-sim/src/testFixtures/java/fr/sncf/osrd/envelope_sim/EnvelopeSimPathBuilder.java
@@ -6,10 +6,10 @@ import com.google.common.collect.ImmutableRangeMap;
 import com.google.common.collect.Range;
 import com.google.common.collect.RangeMap;
 import com.google.common.collect.TreeRangeMap;
-import fr.sncf.osrd.envelope_sim.electrification.Electrified;
-import fr.sncf.osrd.envelope_sim.electrification.NonElectrified;
 import fr.sncf.osrd.path.implementations.EnvelopeSimPath;
 import fr.sncf.osrd.path.interfaces.Electrification;
+import fr.sncf.osrd.path.legacy_objects.electrification.Electrified;
+import fr.sncf.osrd.path.legacy_objects.electrification.NonElectrified;
 import java.util.HashMap;
 
 /**
