+++ core/src/main/java/fr/sncf/osrd/standalone_sim/result/ElectrificationRange.java
@@ -4,10 +4,10 @@ import com.google.common.base.MoreObjects;
 import com.google.common.collect.RangeMap;
 import com.squareup.moshi.Json;
 import com.squareup.moshi.adapters.PolymorphicJsonAdapterFactory;
-import fr.sncf.osrd.envelope_sim.electrification.Electrified;
-import fr.sncf.osrd.envelope_sim.electrification.Neutral;
-import fr.sncf.osrd.envelope_sim.electrification.NonElectrified;
 import fr.sncf.osrd.path.interfaces.Electrification;
+import fr.sncf.osrd.path.legacy_objects.electrification.Electrified;
+import fr.sncf.osrd.path.legacy_objects.electrification.Neutral;
+import fr.sncf.osrd.path.legacy_objects.electrification.NonElectrified;
 import fr.sncf.osrd.train.RollingStock;
 import fr.sncf.osrd.utils.jacoco.ExcludeFromGeneratedCodeCoverage;
 import java.util.ArrayList;
