+++ core/src/main/java/fr/sncf/osrd/train/RollingStock.java
@@ -6,10 +6,10 @@ import com.google.common.collect.RangeMap;
 import com.google.common.collect.TreeRangeMap;
 import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
 import fr.sncf.osrd.envelope_sim.PhysicsRollingStock;
-import fr.sncf.osrd.envelope_sim.electrification.Electrified;
-import fr.sncf.osrd.envelope_sim.electrification.Neutral;
-import fr.sncf.osrd.envelope_sim.electrification.NonElectrified;
 import fr.sncf.osrd.path.interfaces.Electrification;
+import fr.sncf.osrd.path.legacy_objects.electrification.Electrified;
+import fr.sncf.osrd.path.legacy_objects.electrification.Neutral;
+import fr.sncf.osrd.path.legacy_objects.electrification.NonElectrified;
 import fr.sncf.osrd.railjson.schema.rollingstock.Comfort;
 import fr.sncf.osrd.railjson.schema.rollingstock.RJSEtcsBrakeParams;
 import fr.sncf.osrd.railjson.schema.rollingstock.RJSLoadingGaugeType;
